Gestione eccezioni e transazioni in modo più accurato nell'ORM

+++ src/main/java/com/progetto_swe/orm/BookDAO.java
@@ -6,9 +6,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 
 import com.progetto_swe.domain_model.*;
-import com.progetto_swe.orm.database_exception.CRUD_exception;
-import com.progetto_swe.orm.database_exception.DataAccessException;
-import com.progetto_swe.orm.database_exception.DatabaseConnectionException;
+import com.progetto_swe.orm.database_exception.*;
 
 public class BookDAO {
 
@@ -18,36 +16,63 @@ public class BookDAO {
         this.connection = ConnectionManager.getConnection();
     }
 
-    public Book getBook(int code) {
+    public Book getBook(int itemCode) throws IdNotFoundException, DatabaseConnectionException {
         try {
             connection = ConnectionManager.getConnection();
-            String query
-                    = "SELECT * FROM Item I JOIN Book B ON I.code = B.code WHERE I.code = ?";
+            String query = """
+                    SELECT * 
+                    FROM Item I JOIN Book B ON I.code = B.code 
+                    WHERE I.code = ?;
+                    """;
             PreparedStatement ps = connection.prepareStatement(query);
-            ps.setInt(1, code);
+            ps.setInt(1, itemCode);
             ResultSet resultSet = ps.executeQuery();
+
             if(!resultSet.next()) {
-                //System.out.println("There is no book in the database with code = " + code + "!");
-                return null;
+                throw new IdNotFoundException("Errore: Book con itemCode [" + itemCode + "] non presente nel DB.");
             }
-            Book book = new Book(code, resultSet.getString("title"), LocalDate.parse(resultSet.getString("publication_date")), Language.valueOf(resultSet.getString("language")),
-                    Category.valueOf(resultSet.getString("category")), resultSet.getString("link"), resultSet.getString("isbn"),
-                    resultSet.getString("publishing_house"), resultSet.getInt("number_of_pages"), resultSet.getString("authors"));
+
+            Book book = new Book(
+                    itemCode,
+                    resultSet.getString("title"),
+                    LocalDate.parse(resultSet.getString("publication_date")),
+                    Language.valueOf(resultSet.getString("language")),
+                    Category.valueOf(resultSet.getString("category")),
+                    resultSet.getString("link"),
+                    resultSet.getString("isbn"),
+                    resultSet.getString("publishing_house"),
+                    resultSet.getInt("number_of_pages"),
+                    resultSet.getString("authors"));
             PhysicalCopiesDAO physicalCopiesDAO = new PhysicalCopiesDAO();
-            book.setPhysicalCopies(physicalCopiesDAO.getPhysicalCopies(code));
+            book.setPhysicalCopies(physicalCopiesDAO.getPhysicalCopies(itemCode));
             return book;
         } catch (SQLException e) {
-            System.out.println("SQLException: " + e.getMessage());
-            return null;//TODO ipotetica eccezione
+            throw new DatabaseConnectionException(e.getCause().toString());
         }
     }
 
-    public int addBook(String title, String publicationDate, String language, String category, String link, String isbn, String publishingHouse, int numberOfPages, String authors) {
+    public int addBook(String title,
+                       String publicationDate,
+                       String language,
+                       String category,
+                       String link,
+                       String isbn,
+                       String publishingHouse,
+                       int numberOfPages,
+                       String authors,
+                       String storagePlace,
+                       int numberOfCopies,
+                       boolean borrowable)
+            throws IdAlreadyExistsException, DatabaseConnectionException {
 
         connection = ConnectionManager.getConnection();
         try {
             //Creazione Item e Book
-            String query = "INSERT INTO Item (title, publication_date, language, category, link, number_of_pages) VALUES (?, ?, ?, ?, ?, ?);";
+            String query = """
+                    INSERT INTO Item (title, publication_date, language, category, link, number_of_pages) 
+                    VALUES (?, ?, ?, ?, ?, ?) 
+                    RETURNING code;
+                    """;
             PreparedStatement ps = connection.prepareStatement(query, Statement.RETURN_GENERATED_KEYS);
             ps.setString(1, title);
             ps.setDate(2, Date.valueOf(publicationDate));
@@ -56,59 +81,74 @@ public class BookDAO {
             ps.setString(5, link);
             ps.setInt(6, numberOfPages);
 
+            ResultSet resultSet = ps.executeQuery();
+            int itemCode = resultSet.getInt("code");
 
-            ps.executeUpdate();
-
-            ResultSet generatedKeys = ps.getGeneratedKeys();
-            if (!generatedKeys.next()) {
-                throw new CRUD_exception("Error executing insert!", null);
-            }
-            int code = generatedKeys.getInt(1);
-
-            query
-                    = "INSERT INTO Book (code, isbn, publishing_house, authors)"
-                    + " VALUES (?, ?, ?, ?);";
+            query = """
+                    INSERT INTO Book (code, isbn, publishing_house, authors) 
+                    VALUES (?, ?, ?, ?);
+                    """;
             ps = connection.prepareStatement(query);
-            ps.setInt(1, code);
+            ps.setInt(1, itemCode);
             ps.setString(2, isbn);
             ps.setString(3, publishingHouse);
             ps.setString(4, authors);
-
             ps.executeUpdate();
 
-            return code;
+            PhysicalCopiesDAO physicalCopiesDAO = new PhysicalCopiesDAO();
+            physicalCopiesDAO.addPhysicalCopies(itemCode, storagePlace, numberOfCopies, borrowable);
+            return itemCode;
         } catch (SQLException e) {
-            System.out.println("SQLException: " + e.getMessage());
-            return -1;//TODO ipotetica eccezione
+            throw new DatabaseConnectionException(e.getCause().toString());
         }
     }
 
-    public void removeBook(int code) {
-
+    public void removeBook(int itemCode) throws IdNotFoundException, ConstrainViolationException, DatabaseConnectionException {
+//TODO gestire transazioni dentro le classi DAO
         connection = ConnectionManager.getConnection();
+        ConnectionManager.closeAutoCommit();
         try {
             String query = "DELETE FROM Book WHERE code = ?;";
             PreparedStatement ps = connection.prepareStatement(query);
-            ps.setInt(1, code);
-            ps.executeUpdate();
+            ps.setInt(1, itemCode);
+
+            if(ps.executeUpdate() != 1) {
+                ConnectionManager.rollback();
+                throw new IdNotFoundException("Errore: Il Book con itemCode [" + itemCode + "] non è presente nel DB.");
+            }
 
             query = "DELETE FROM Item WHERE code = ?;";
-
-
             ps = connection.prepareStatement(query);
-            ps.setInt(1, code);
+            ps.setInt(1, itemCode);
 
-            //return ps.executeUpdate() != 0; TODO controllo se tupla non esistente restituisce 0 oppure lancia eccezione
-
+            if(ps.executeUpdate() != 1) {
+                ConnectionManager.rollback();
+                throw new IdNotFoundException("Errore: L'Item con ItemCode [" + itemCode + "] che vuoi rimuovere non è un Book. [problema su chiamata del metodo]");
+            }
+            ConnectionManager.commit();
         } catch (SQLException e) {
-            System.out.println("SQLException: " + e.getMessage());
-            //TODO ipotetica eccezione
+            ConnectionManager.rollback();
+            if(e.getSQLState().equals("23503")){
+                throw new ConstrainViolationException("Errore: Book con itemCode [" + itemCode + "] non può essere eliminato perché sono ancora presenti Copie/Prenotazioni/Prestiti. [problema del programma controllare logica di cancellazione elemento]");
+            }
+            throw new DatabaseConnectionException(e.getCause().toString());
         }
     }
 
-    public void updateBook(int originalItemCode, String title, String publicationDate, String language, String category, String link, String isbn,
-                              String publishingHouse, String authors) {
+    public void updateBook(int originalItemCode,
+                           String title,
+                           String publicationDate,
+                           String language,
+                           String category,
+                           String link,
+                           String isbn,
+                           String publishingHouse,
+                           String authors,
+                           String storagePlace,
+                           int newNumberOfCopies,
+                           boolean borrowable) throws IdNotFoundException, ConstrainViolationException, DatabaseConnectionException{
         connection = ConnectionManager.getConnection();
+        ConnectionManager.closeAutoCommit();
         try {
             String query
                     = "UPDATE Item SET title = ?, publication_date = ?, language = ?, category = ?, link = ? WHERE code = ?;";
@@ -120,29 +160,32 @@ public class BookDAO {
             ps.setString(5, link);
             ps.setInt(6, originalItemCode);
 
-            //if(ps.executeUpdate() == 0){
-            //    return false;
-            //} TODO controllo se tupla non esistente restituisce 0 oppure lancia eccezione
+            if(ps.executeUpdate() != 1) {
+                ConnectionManager.rollback();
+                throw new IdNotFoundException("Errore: Book con ItemCode [" + originalItemCode + "] non presente nel DB.");
+            }
 
             query = "UPDATE Book SET isbn = ?, publishing_house = ?, authors = ? WHERE code = ?;";
-
             ps = connection.prepareStatement(query);
             ps.setString(1, isbn);
             ps.setString(2, publishingHouse);
             ps.setString(3, authors);
             ps.setInt(4, originalItemCode);
 
-            //return ps.executeUpdate() != 0;TODO controllo se tupla non esistente restituisce 0 oppure lancia eccezione
+            if(ps.executeUpdate() != 1) {
+                ConnectionManager.rollback();
+                throw new IdNotFoundException("Errore: Item con ItemCode [" + originalItemCode + "] che cuoi aggiornare non è un Book.");
+            }
+            PhysicalCopiesDAO physicalCopiesDAO = new PhysicalCopiesDAO();
+            physicalCopiesDAO.updatePhysicalCopies(originalItemCode, storagePlace, newNumberOfCopies, borrowable);
         } catch (SQLException e) {
-            System.out.println("SQLException: " + e.getMessage());
-            //TODO ipotetica eccezione
+            throw new DatabaseConnectionException(e.getCause().toString());
         }
     }
 
 
-    public ArrayList<Book> getAllBooks() {
+    public ArrayList<Book> getAllBooks() throws DatabaseConnectionException {
         ArrayList<Book> books = new ArrayList<>();
-        HashMap<Library, PhysicalCopies> physicalCopies;
         connection = ConnectionManager.getConnection();
         try {
             //tutti i book
@@ -164,24 +207,12 @@ public class BookDAO {
             }
 
             for(Book book : books){
-                String query_2 = "SELECT * FROM physical_copies P WHERE P.code = ?;";
-                ps = connection.prepareStatement(query_2);
-                ps.setInt(1, book.getCode());
-                ResultSet copiesSet = ps.executeQuery();
-                physicalCopies = new HashMap<>();
-                while (copiesSet.next()) {
-                    physicalCopies.put(
-                            Library.valueOf(copiesSet.getString("storage_place")),
-                            new PhysicalCopies(copiesSet.getInt("number_of_copies"),
-                                    copiesSet.getInt("number_of_available_copies"),
-                                    copiesSet.getBoolean("borrowable")));
-                }
-                book.setPhysicalCopies(physicalCopies);
+                PhysicalCopiesDAO physicalCopiesDAO = new PhysicalCopiesDAO();
+                book.setPhysicalCopies(physicalCopiesDAO.getPhysicalCopies(book.getCode()));
             }
             return books;
         } catch (SQLException e) {
-            System.out.println("SQLException: " + e.getMessage());
-            return null; //ipotetica eccezione
+            throw new DatabaseConnectionException(e.getCause().toString());
         }
     }
